Edge MC simulation working fine, calculating delay of the two possible delays at each edge(when possible). Also create RE data structures for RE MC sim and validated it. Next step is to do MC sim with RE.

+++ multiple_paths/completeNetlist.cpp
@@ -115,7 +115,7 @@ void update_currentFabric() // resets the current global variables to the origin
 void set_number_of_bitstreams_ohne_feedback()
 {
 
-	bitstreamsOhneFeedback = testedPaths.size();
+	bitstreamsOhneFeedback = (int)testedPaths.size();
 
 }
